Fix function annotation for dataflow tasks
We introduce a callable finalization struct that is invoked
when posting the final function for execution.
An overload for get_function_annotation for the struct
is provided and it returns the correct type.

Works for all dataflow except launch::async policy as this
does not create a new task and therefore we do not pick up name
changes.